fixed windows make files and VS2012-specific compilation issues

+++ src/share/vm/graal/graalCompiler.cpp
@@ -49,8 +49,8 @@ void GraalCompiler::initialize() {
   JavaThread* THREAD = JavaThread::current();
   TRACE_graal_1("GraalCompiler::initialize");
 
-  unsigned long heap_end = (long) Universe::heap()->reserved_region().end();
-  unsigned long allocation_end = heap_end + 16l * 1024 * 1024 * 1024;
+  uintptr_t heap_end = (uintptr_t) Universe::heap()->reserved_region().end();
+  uintptr_t allocation_end = heap_end + ((uintptr_t)16) * 1024 * 1024 * 1024;
   guarantee(heap_end < allocation_end, "heap end too close to end of address space (might lead to erroneous TLAB allocations)");
   NOT_LP64(error("check TLAB allocation code for address space conflicts"));
